% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiuc_enrollment_data.R
\docType{data}
\name{curriculum_and_class_level}
\alias{curriculum_and_class_level}
\alias{fall_enrolled_ranking}
\alias{spring_enrolled_ranking}
\alias{summer_enrolled_ranking}
\title{Enrollment based on Curriculum and Class Level Data from Fall 2004 - Spring 2017}
\format{
There are three different data sets documented simultaneously within
this file.
\itemize{
\item \code{fall_curriculum_class}: 8590 observations and 23 variables.
\item \code{spring_curriculum_class}: 8760 observations and 23 variables.
\item \code{summer_curriculum_class}: 6369 observations and 23 variables.
}

with the format of:
\itemize{
\item \code{term_code}
\itemize{
\item Factor
\item ID for the Current Semester
}
\item \code{college}
\itemize{
\item Factor
\item See the \href{http://registrar.illinois.edu}{Office of the Registrar}
}
\item \code{department}
\itemize{
\item Factor
\item Department Code
}
\item \code{degree}
\itemize{
\item Factor
\item See \url{http://catalog.illinois.edu/degree-programs/}
}
\item \code{major_code}
\itemize{
\item Factor
\item Major number
}
\item \code{major_name}
\itemize{
\item Factor
\item Name of the Major
}
\item \code{concentration_code}
\itemize{
\item Factor
\item ID of Major Concentration
}
\item \code{concentration_name}
\itemize{
\item Factor
\item Name of the Major Concentration
}
\item \code{freshman}
\itemize{
\item Integer
\item Count of Enrolled Freshman
}
\item \code{sophomore}
\itemize{
\item Integer
\item Count of Enrolled Sophomores
}
\item \code{junior}
\itemize{
\item Integer
\item Count of Enrolled Juniors
}
\item \code{senior}
\itemize{
\item Integer
\item Count of Enrolled Seniors
}
\item \code{nondegree_undergrad}
\itemize{
\item Integer
\item Count of Non-degree Program Undergraduates
}
\item \code{total_undergrad}
\itemize{
\item Integer
\item Count of All Enrolled Undergraduates
}
\item \code{grad_master}
\itemize{
\item Integer
\item Count of Enrolled Master Graduate Students
}
\item \code{grad_doctoral}
\itemize{
\item Integer
\item Count of Enrolled Doctoral Graduate Students
}
\item \code{nondegree_grad}
\itemize{
\item Integer
\item Count of Enrolled Non-degree Program Graduate Students
}
\item \code{total_grad}
\itemize{
\item Integer
\item Count of All Enrolled Graduate Students
}
\item \code{professional}
\itemize{
\item Integer
\item Count of All Professional Students
}
\item \code{overall_total}
\itemize{
\item Integer
\item Count of All Enrolled Students
}
\item \code{academic_program_code}
\itemize{
\item Factor
\item University ID Code for Program of Study
\item e.g. 10KL0010BS, where BS represents degre
}
\item \code{year}
\itemize{
\item Factor
\item 2005 - 2016
}
\item \code{term}
\itemize{
\item Factor
\item Fall, Spring, or Summer
}
}

An object of class \code{data.frame} with 8760 rows and 23 columns.

An object of class \code{data.frame} with 6369 rows and 23 columns.
}
\usage{
fall_enrolled_ranking

spring_enrolled_ranking

summer_enrolled_ranking
}
\description{
This report uses the final enrollment statistics from the official tenth day
of on-campus terms. The reports are sorted by college, department code, and
major code.
}
\details{
Only students taking at least one on-campus, credit-bearing class
are included in these reports. The following categories of students are
excluded:
\itemize{
\item auditors (students taking only non-credit classes);
\item students taking only extramural or off-campus classes;
\item Medical Scholars taking no on-campus, non-MSP classes.
\itemize{
\item Note: Illini Center MBA students are included in these statistics.
}
}
}
\keyword{datasets}
