% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachers-ranked-excellent-data.R
\docType{data}
\name{excellent_teachers}
\alias{excellent_teachers}
\title{Teachers Ranked as Excellent from Fall 1993 through Summer 2017}
\format{
A \code{data.frame} with 85,778 observations and 9 variables:
\itemize{
\item \code{term}: Two letter semester code (\code{sp}, \code{su}, \code{fa}, or \code{wi}) followed by a four digit year.
\itemize{
\item Examples: \code{sp2017}, \code{fa2013}, \code{su2012}.
}
\item \code{semester}: Two letter semester code
\itemize{
\item Examples: \code{sp}, \code{su}, \code{fa}, or \code{wi}
}
\item \code{year}: Four digit year.
\itemize{
\item Examples: \code{2017}, \code{2013}, \code{2012}.
}
\item \code{unit}: The CITL-supplied headers for the unit teaching the course.
\itemize{
\item Examples: \code{ACCOUNTANCY}, \verb{SOCIAL WORK}, \code{LINGUISTICS}, \verb{NUCLEAR, PLASMA & RAD. ENGR.}
}
\item \code{lname}: The last name of the teacher.
\itemize{
\item Examples: \code{FAGEN-ULMSCHNEIDER}, \code{FLANAGAN}, \code{FLECK}
}
\item \code{fname}: The first letter of the first name of the teacher.
\itemize{
\item Examples: \code{W}, \code{K}, \code{M}
}
\item \code{role}: \code{Instructor} or \code{TA}
\item \code{ranking}: \code{Excellent} or \code{Outstanding}
\item \code{course}: The course the teacher was ranked as excellent.  If no course is given, the \code{course} field is set to \code{NA} (this includes cases when the raw data lists the course as \code{0}, \code{000}, or \code{999}).
\itemize{
\item Examples: \code{199}, \code{225}, \code{560}, \code{NA}
}
}
}
\source{
Wade Fagen-Ulmschneider

Teaching Assistant Professor, Computer Science

University of Illinois at Urbana-Champaign

\url{http://waf.cs.illinois.edu/}
}
\usage{
excellent_teachers
}
\description{
A collection of all "Teachers Ranked as Excellent" at UIUC from Fall 1993
through Summer 2017.
}
\details{
At the University of Illinois at Urbana-Champaign, every semester students
complete an \href{http://citl.illinois.edu/citl-101/measurement-evaluation/teaching-evaluation/teaching-evaluations-(ices)}{Instructor and Course Evaluation (ICES)} questionnaire for each
course they are enrolled.  The \href{http://citl.illinois.edu/}{Center for Innovation in Teaching and Learning (CITL)} compiles this data into the list of "Teachers Ranked as Excellent"
and publishes the names of those teaches in the local newspaper and
\href{http://citl.illinois.edu/citl-101/measurement-evaluation/teaching-evaluation/teaching-evaluations-(ices)/teachers-ranked-as-excellent}{online as a PDF}
}
\references{
\url{https://github.com/wadefagen/Teachers-Ranked-As-Excellent-UIUC}
}
\keyword{datasets}
