% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiuc_grades_data.R
\docType{data}
\name{grade_dist}
\alias{grade_dist}
\title{Grade Distribution Data from Summer 2005 - Summer 2017}
\format{
A data.frame with 124155 and 25 variables with the format of:
\itemize{
\item \code{crn}
\itemize{
\item Course Registration Number
}
\item \code{subject}
\itemize{
\item Department course was offered in
}
\item \code{course}
\itemize{
\item Course Number
}
\item \code{course_title}
\itemize{
\item Name of the Course
}
\item \code{course_section}
\itemize{
\item Lecture, Discussion, or Online Section ID
}
\item \code{sched_type}
\itemize{
\item Lecture, Discussion, Online
}
\item \code{a_plus}
\itemize{
\item Number of Students Receiving an A+
\item Weighted as 4.00
}
\item \code{a}
\itemize{
\item Number of Students Receiving an A
\item Weighted as 4.00
}
\item \code{a_minus}
\itemize{
\item Number of Students Receiving an A-
\item Weighted as 3.67
}
\item \code{b_plus}
\itemize{
\item Number of Students Receiving an B+
\item Weighted as 3.33
}
\item \code{b}
\itemize{
\item Number of Students Receiving an B
\item Weighted as 3.00
}
\item \code{b_minus}
\itemize{
\item Number of Students Receiving an B-
\item Weighted as 2.67
}
\item \code{c_plus}
\itemize{
\item Number of Students Receiving an C+
\item Weighted as 2.33
}
\item \code{c}
\itemize{
\item Number of Students Receiving an C
\item Weighted as 2.00
}
\item \code{c_minus}
\itemize{
\item Number of Students Receiving an C-
\item Weighted as 1.67
}
\item \code{D_plus}
\itemize{
\item Number of Students Receiving an D+
\item Weighted as 1.33
}
\item \code{d}
\itemize{
\item Number of Students Receiving an D
\item Weighted as 1.00
}
\item \code{d_minus}
\itemize{
\item Number of Students Receiving an D-
\item Weighted as 0.67
}
\item \code{f}
\itemize{
\item Number of Students Receiving an F
\item Weighted as 0.00
}
\item \code{w}
\itemize{
\item Number of Students Receiving an W (Withdrawal)
}
\item \code{average_grade}
\itemize{
\item The average GPA for the course
}
\item \code{semester}
\itemize{
\item Name of term: Fall, Winter, Spring, or Summer
}
\item \code{year}
\itemize{
\item Year semester takes place in.
}
}
}
\usage{
grade_dist
}
\description{
Grade distributions by letter grade for every class and instructor from
Summer 2005 to Summer 2017 at the University of Illinois at Urbana-Champaign
(UIUC).
}
\details{
As the R data file type allows for editing of the included data, we ask that
prior to using the files you accept responsibility for accurate dissemination
of this information and agree that no manipulation of the data that results
in falsification of the information will occur.

Missing values contained within the data set are due to the
grade distributions not being displayed when a section has low enrollment or
when all students in the class have the same grade.
Because of the low enrollment in those classes or because all the students
in a class received the same grade, the grade data could identify a student.
Therefore, this data has been omitted to satisfy Family Educational Rights
and Privacy Act (FERPA) (20 U.S.C. §1232g).

Please see the \href{https://registrar.illinois.edu/explanation-of-grades}{Grade Scale}
for details as to how the average GPA was calculated.
}
\examples{

# Load Fall Data
data("grade_dist", package="uiucdata")
}
\keyword{datasets}
