% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiucdata_pkg.R
\docType{package}
\name{uiucdata-package}
\alias{uiucdata}
\alias{uiucdata-package}
\title{uiucdata: University of Illinois at Urbana-Champaign Data Package}
\description{
Contained within this package are University of Illinois at Urbana-Champaign
specific data sets. These data sets range from university financial transactions
to enrollment numbers.
}
\details{
Data presented here has been obtained either through a FOIA request or
by talking with the respective representative.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
